+++ src/main/java/com/gmail/ilerofv/entity/Transaction.java
@@ -1,6 +1,5 @@
 package com.gmail.ilerofv.entity;
 
-import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -8,6 +7,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Getter
@@ -19,9 +19,9 @@ import java.util.Date;
 public class Transaction {
     private String id;
     private Date date;
-    private double amount;
+    private BigDecimal amount;
     private String merchant;
     private TransactionType transactionType;
     private String related = "";
-    
+
 }
